do not remove dev packages; should save julia
Julia has been getting uninstalled because it seems to depend on a particular
dev package: librmath-julia-dev. For now, I am avoiding uninstalling the dev
packages because some of them seem necessary. I will see how this affects
image size, and if it has a serious detrimental effect, i will try
cherry-picking the not-uninstalled-dev packages while removing the rest.

Also, since this should fix julia again, add the julia test back in, along with
a couple of 'todo' comments on future tests that should be added.