install vdt w/ datagrid; just skip julia uninstall
vdt-gsi-openssh-client apparently comes with the ligo datagrid repositories;
an error was coming up when i mistakenly installed it after only adding the
globus repo.

the only dev package that was obviously necessary was the julia dev package.
set the uninstall script to skip to leave anything with the name 'julia' in it
alone. this allows for the removal of other dev packages.